Break out ExternalNetworkIcon into its own file
